refactor(mf): runner - scss into css
