--- micro-frontends/src/mf-runner/src/ui/modules/controller/controller.module.scss
@@ -1,6 +1,0 @@
-.controller {
-	display: flex;
-	column-gap: 5px;
-	justify-content: center;
-	margin-top: 15px;
-}
