+++ micro-frontends/src/mf-runner/src/ui/modules/description/description.module.css
@@ -1,0 +1,5 @@
+.description {
+	display: flex;
+	column-gap: var(--ds-spacing-xl);
+	margin-bottom: var(--ds-spacing-sm);
+}
