--- micro-frontends/src/mf-runner/src/ui/modules/description/description.module.scss
@@ -1,5 +1,0 @@
-.description {
-	display: flex;
-	column-gap: 10px;
-	margin-bottom: 10px;
-}
