+++ micro-frontends/src/mf-runner/src/ui/modules/modal/modal.module.css
@@ -1,0 +1,3 @@
+.wrapper {
+	padding: var(--ds-spacing-lg);
+}
