--- micro-frontends/src/mf-runner/src/ui/modules/modal/modal.module.scss
@@ -1,3 +1,0 @@
-.wrapper {
-    padding: 20px;
-}
